backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-Session
Werkzeug
python-dotenv
psycopg2-binary
Pillow
bcrypt
requests
gunicorn
SQLAlchemy
cachelib
msgspec